requirements.txt
fastapi
uvicorn[standard]
google-api-python-client
earthengine-api
pydantic[email]
python-dotenv
earthengine-api==1.6.6
matplotlib==3.10.6
geemap==0.36.2
pandas==2.3.2
mysql-connector-python
requests
sqlalchemy
schemas
cryptography
python-decouple
python-multiple
geopy